helper infowindow button fix

+++ app/routes/main.py
@@ -44,9 +44,12 @@ def helper():
         longitude = float(request.form['longitude'])
         latitude = float(request.form['latitude'])
         phone = request.form['phone']
+
         user = User.query.filter(and_(User.longitude==longitude, User.latitude==latitude, User.phone==phone)).first()
-        db.session.delete(user)
-        db.session.commit()
+        print(request.form, user)
+        if user is not None:
+            db.session.delete(user)
+            db.session.commit()
 
         # emit('remove helpee', {data: 'test'}, broadcast=True)
         return jsonify(success=True), 200
